Issue #715: ClassificationListコンポーネントのカスタムフック作成
